refactor all models into 3 concerns
