+++ model/src/applications/command/mod.rs
@@ -1,7 +1,5 @@
-mod callback_data;
 mod option;
 
-pub use callback_data::CommandCallbackData;
 pub use option::{
     BaseCommandOptionData, ChoiceCommandOptionData, CommandOption, CommandOptionChoice,
     CommandOptionType, OptionsCommandOptionData,
@@ -10,7 +8,7 @@ pub use option::{
 use crate::id::{ApplicationId, CommandId};
 use serde::{Deserialize, Serialize};
 
-/// An executable command by a user in a server.
+/// Data sent to discord to create a command.
 ///
 /// Refer to [the discord docs] for more information.
 ///
@@ -28,58 +26,4 @@ pub struct Command {
     pub description: String,
     #[serde(default)]
     pub options: Vec<CommandOption>,
-}
-
-/// Data received when an [`ApplicationCommand`] interaction is executed.
-///
-/// Refer to [the discord docs] for more information.
-///
-/// [`ApplicationCommand`]: crate::applications::interaction::Interaction::ApplicationCommand
-/// [the discord docs]: https://discord.com/developers/docs/interactions/slash-commands#interaction-applicationcommandinteractiondata
-#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize, Default)]
-pub struct CommandData {
-    pub id: CommandId,
-    pub name: String,
-    #[serde(default)]
-    pub options: Vec<CommandDataOption>,
-}
-
-/// Data received when a user fills in a command option.
-///
-/// Note: user, channel, and role option types will be returned as a String
-/// option here.
-///
-/// Refer to [the discord docs] for more information.
-///
-/// [the discord docs]: https://discord.com/developers/docs/interactions/slash-commands#interaction-applicationcommandinteractiondataoption
-#[derive(Clone, Debug, Deserialize, Eq, Hash, PartialEq, Serialize)]
-#[serde(untagged)]
-pub enum CommandDataOption {
-    String {
-        name: String,
-        value: String,
-    },
-    Integer {
-        name: String,
-        value: i64,
-    },
-    Boolean {
-        name: String,
-        value: bool,
-    },
-    SubCommand {
-        name: String,
-        options: Vec<CommandDataOption>,
-    },
-}
-
-impl CommandDataOption {
-    pub fn kind(&self) -> &'static str {
-        match self {
-            CommandDataOption::String { .. } => "String",
-            CommandDataOption::Integer { .. } => "Integer",
-            CommandDataOption::Boolean { .. } => "Boolean",
-            CommandDataOption::SubCommand { .. } => "SubCommand",
-        }
-    }
 }
