+++ model/src/applications/mod.rs
@@ -1,2 +1,6 @@
+/// Used for building commands to send to Discord.
 pub mod command;
+/// Used when recieving interactions through gateway or webhooks.
 pub mod interaction;
+/// Used when responding to interactions.
+pub mod response;
