+++ model/src/applications/response/mod.rs
@@ -1,9 +1,11 @@
-use std::convert::{TryFrom, TryInto};
-use std::fmt::{Display, Formatter, Result as FmtResult};
+mod callback_data;
 
-use crate::applications::command::CommandCallbackData;
+pub use callback_data::CommandCallbackData;
+
 use serde::{Deserialize, Deserializer, Serialize, Serializer};
 use serde_repr::{Deserialize_repr, Serialize_repr};
+use std::convert::{TryFrom, TryInto};
+use std::fmt::{Display, Formatter, Result as FmtResult};
 
 /// Payload used for responding to an interaction.
 ///
